Create virtual containers for entities
The subjects(), predicates() and objects() methods now return "virtual"
containers that have a forEach() function to asynchronously loop over all
entries in the container.

The instances of EntityIterator have now been moved to a separate
iterator() function on these virtual containers, which returns the same
iterator that the entity functions used to return.